Create CHEGG 01
The materials in the transport company are loaded with materials of different weights. The program calculates how many materials and how many total materials are loaded when the truck is full. The program stops when -1 is entered.

Program Output:

Enter 1. weight of the material: 20

Enter 2. weight of the material: 30

Enter 3. weight of the material: 45

Enter 4. weight of the material: 10

Enter 5. weight of the material: -1

4 Material Loaded on Truck

Total Weight of Materials: 105

Programming language is python. Thank you for your efforth and knowledge